@@ -1,32 +1,30 @@
-ruby '2.4.5'
+ruby '2.6.5'
 
 source 'https://rubygems.org'
 
-gem 'rails'
-
 gem 'active_model_serializers'
-gem 'bcrypt-ruby'
+gem 'bcrypt'
 gem 'bootstrap-sass'
 gem 'browser-timezone-rails'
 gem 'carrierwave'
 gem 'coffee-rails'
-gem 'fog'
 gem 'font-awesome-rails'
 gem 'google-search'
 gem 'haml-rails'
 gem 'jbuilder'
 gem 'jquery-rails'
 gem 'jquery-ui-sass-rails'
-gem 'pg', '~> 0.18'
+gem 'pg'
+gem 'rails'
 gem 'redcarpet'
-gem 'rmagick', :require => 'RMagick'
+gem 'rmagick', :require => 'rmagick'
 gem 'sass-rails'
+gem 'therubyracer'
 gem 'turbolinks'
 gem 'uglifier'
 gem 'will_paginate'
 
 group :development do
-  gem 'spring'
   gem 'thin'
   gem 'pry-rails'
 end
@@ -42,8 +40,4 @@ group :test do
   gem 'capybara'
   gem 'simplecov'
   gem 'coveralls', require: false
-  gem 'guard-rails'
-  gem 'guard-rspec'
-  gem 'guard-spork'
-  gem 'spork-rails', :github => 'sporkrb/spork-rails'
 end
