@@ -1,4 +1,4 @@
-class CreateFamilies < ActiveRecord::Migration
+class CreateFamilies < ActiveRecord::Migration[4.2]
   def change
     create_table :families do |t|
       t.string :name
