@@ -1,4 +1,4 @@
-class CreateWishItems < ActiveRecord::Migration
+class CreateWishItems < ActiveRecord::Migration[4.2]
   def change
     create_table :wish_items do |t|
       t.integer :wish_list_id
