@@ -1,4 +1,4 @@
-class CreateDashboards < ActiveRecord::Migration
+class CreateDashboards < ActiveRecord::Migration[4.2]
   def change
     create_table :dashboards do |t|
       t.string :owner_type
