@@ -1,4 +1,4 @@
-class AddFamilyMemberships < ActiveRecord::Migration
+class AddFamilyMemberships < ActiveRecord::Migration[4.2]
   def change
     create_table :family_memberships do |t|
       t.integer :user_id
