@@ -1,4 +1,4 @@
-class AddOwnerIdToFamily < ActiveRecord::Migration
+class AddOwnerIdToFamily < ActiveRecord::Migration[4.2]
   def change
     add_column :families, :owner_id, :integer
   end
