@@ -1,4 +1,4 @@
-class AddUrlToWishItem < ActiveRecord::Migration
+class AddUrlToWishItem < ActiveRecord::Migration[4.2]
   def change
     add_column :wish_items, :url, :string
   end
